# Fix inefficiencies in auto-linking code